requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0

